Crate trait_set

This crate provide support for trait aliases: a feature that is already supported by Rust compiler, but is not stable yet.

The idea is simple: combine group of traits under a single name. The simplest example will be:

use trait_set::trait_set;

trait_set! {
    pub trait ThreadSafe = Send + Sync;
}

Macro trait_set displayed here is the main entity of the crate: it allows declaring multiple trait aliases, each of them is represented as

[visibility] trait [AliasName][<generics>] = [Element1] + [Element2] + ... + [ElementN];

For more details, see the trait_set macro documentation.

Macros

trait_set

Creates an alias for set of traits.